memcg: change memcg->private_id to unsigned short
The private_id field is allocated via xa_alloc() bounded by
MEM_CGROUP_ID_MAX, which is (1 << MEM_CGROUP_ID_SHIFT) - 1 = 65535.
This fits in an unsigned short. So change the field's type to reflect
its actual range.

Since xa_alloc() requires an unsigned int *, use a temporary local
variable for the allocation and assign it to memcg->private_id
afterwards.

+++ include/linux/memcontrol.h
@@ -325,7 +325,7 @@ struct mem_cgroup {
 #endif
 
 	/* Private memcg ID. Used to ID objects that outlive the cgroup */
-	int private_id;
+	unsigned short private_id;
 
 	int kmemcg_id;
